+++ app/(features)/(user)/myactivities/page.tsx
@@ -1,18 +1,27 @@
+"use client"; // 클라이언트 컴포넌트로 지정
+
+import DropDownMenu from "@/components/DropDown/ActivityEditDelete";
+import ReservationCard from "@/components/ReservationCard";
+import { ActivitiesData, ActivityList } from "@/types/activities.type";
 import getMyActivities from "@api/MyActivities/getMyActivities";
 import DefaultButton from "@button/DefaultButton";
 import Link from "next/link";
-import MyActivitiesClientBody from "./_components/MyActivitiesClientBody";
+import { useEffect, useState } from "react";
 
-export default async function MyActivities() {
-  // useEffect(() => {
-  //   async function fetchData() {
-  //     const data = await getMyActivities({ size: 20 });
-  //     setInitialData(data);
-  //   }
-  //   fetchData();
-  // }, []);
+export default function MyActivities() {
+  const [initialData, setInitialData] = useState<ActivitiesData | null>(null);
 
-  const data = await getMyActivities({ size: 20 });
+  useEffect(() => {
+    async function fetchData() {
+      const data = await getMyActivities({ size: 20 });
+      setInitialData(data);
+    }
+    fetchData();
+  }, []);
+
+  if (!initialData) {
+    return <div>Loading...</div>;
+  }
 
   return (
     <div>
@@ -27,7 +36,26 @@ export default async function MyActivities() {
           </DefaultButton>
         </Link>
       </div>
-      <MyActivitiesClientBody initialData={data} />
+      <div className="space-y-2 md:space-y-4 xl:space-y-6">
+        {initialData.activities.length === 0 ? (
+          <div>No activities found.</div>
+        ) : (
+          initialData.activities.map((activity: ActivityList) => (
+            <ReservationCard
+              key={activity.id}
+              reservations={[activity]}
+              getImageUrl={(activity: ActivityList) => activity.bannerImageUrl || ""}
+              getTitle={(activity: ActivityList) => activity.title}
+              maxTitleLength={18}
+              getRating={(activity: ActivityList) => activity.rating}
+              getReviewCount={(activity: ActivityList) => activity.reviewCount}
+              getPrice={(activity: ActivityList) => activity.price}
+            >
+              <DropDownMenu activityId={activity.id} />
+            </ReservationCard>
+          ))
+        )}
+      </div>
     </div>
   );
 }
